feat: Identity v2 (#705)

+++ packages/wasm-dpp/src/errors/consensus/basic/identity/duplicated_identity_public_key_id_error.rs
@@ -21,7 +21,7 @@ impl DuplicatedIdentityPublicKeyIdErrorWasm {
         self.inner
             .duplicated_ids()
             .iter()
-            .map(|id| JsValue::from(*id as u32))
+            .map(|id| JsValue::from(*id))
             .collect()
     }
